+++ Krowi_AchievementFilter.lua
@@ -54,6 +54,8 @@ local function LoadKrowi_AchievementFilter()
     addon.BrowsingHistory:Load();
 
     addon.TooltipData.Load(); -- Might be moved to PLAYER_LOGIN event but easier for testing on every /reload
+
+    addon.Diagnostics.TaintProbe.Load();
 end
 
 local function LoadBlizzard_AchievementUI()